requirements.txt
anthropic>=0.7.0
flask>=3.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
werkzeug>=3.0.0
